Moved original App top level component to new folder 'TestApp'. Defined outline of roles and pages to create.
