+++ package.json
@@ -2,6 +2,7 @@
   "name": "volunteer-app",
   "version": "1.0.0",
   "main": "src/app/index.ts",
+  "type": "module",
   "scripts": {
     "start": "expo start",
     "android": "expo start --android",
