+++ src/app/index.ts
@@ -2,7 +2,9 @@ import { registerRootComponent } from 'expo';
 
 // We import the top-level component App here and use it as our root, where everything inside of 
 // App's return statement will be rendered
-import App from '../components/App/App';
+
+// change this to App instead of OriginalApp if you want to test
+import App from '../components/TestApp/App';
 
 // registerRootComponent calls AppRegistry.registerComponent('main', () => App);
 // It also ensures that whether you load the app in Expo Go or in a native build,
